+++ CHANGELOG.md
@@ -3,6 +3,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
+## [v1.2b5-1.16.5] - 2021-07-22
+### Added
+- Re-added flowy ocelot tail model
+- Re-added curly cat tail model
+
 ## [v1.2b4-1.16.5] - 2021-07-22
 ### Added
 - Re-added wiggly iron golem nose model